Move meta/card
