+++ frontend/src/metabase/modes/components/TimeseriesFilterWidget.jsx
@@ -8,7 +8,7 @@ import PopoverWithTrigger from "metabase/components/PopoverWithTrigger";
 import SelectButton from "metabase/core/components/SelectButton";
 import Button from "metabase/core/components/Button";
 
-import * as Card from "metabase/meta/Card";
+import * as Card from "metabase-lib/lib/queries/utils/card";
 import { generateTimeFilterValuesDescriptions } from "metabase-lib/lib/queries/utils/query-time";
 import * as Query from "metabase-lib/lib/queries/utils/query";
 import * as Filter from "metabase-lib/lib/queries/utils/filter";
